Reset tty to a "reasonable" state before processing stdin.
In which we go on a short journey of adventure into the land of POSIX
terminal handling to reset canonical mode, etc., since the Pony runtime
puts it into something like raw mode.

Added a pony_unix C helper library to manipulate struct termios.

Added a unix.pony module holding the necessary Unistd and Termios
functionality.

Reset the terminal to it's (experimentally determined, on this Arch
Linux box) default state.